refactor: improve type safety and add BLE API aliases
- Add type annotations for config object and timeout parameters
- Fix potential None access in export_config when myinfo is None
- Treat timestamp value 0 as None in time formatting methods
- Add CamelCase aliases for BLE interface methods for API compatibility
- Improve test mocking for set-owner and set-ham commands
- Fix _FallbackDotMap to support attribute setting and deletion
- Apply consistent docstring formatting across modules
- Rename client_factory to resolved_factory for clarity in discovery

+++ meshtastic/stream_interface.py
@@ -31,7 +31,7 @@ class StreamInterface(MeshInterface):
         noProto: bool = False,
         connectNow: bool = True,
         noNodes: bool = False,
-        timeout: int = 300,
+        timeout: float = 300.0,
     ) -> None:
         """
         Initialize the StreamInterface, prepare its reader thread, and optionally open and configure the underlying stream connection.
@@ -42,7 +42,7 @@ class StreamInterface(MeshInterface):
             noProto (bool): If True, skip protocol-specific startup and allow using this class without a concrete stream implementation.
             connectNow (bool): If True, call connect() after initialization and, unless `noProto` is True, wait for protocol configuration.
             noNodes (bool): Passed to the MeshInterface initializer to control node discovery behavior.
-            timeout (int): Seconds to wait for replies and configuration operations.
+            timeout (float): Seconds to wait for replies and configuration operations.
 
         Raises
         ------
